+++ tests/unit/sqlite/test_common.cpp
@@ -636,7 +636,7 @@ public:
             std::vector<resultType> true_rows = {{"Alice", "Apple", 500}, {"Bob", "TV", 900}};
             int                     idx       = 0;
             ilias_for_await(auto &row, result.range()) {
-                if (idx >= true_rows.size()) {
+                if (idx >= (int)true_rows.size()) {
                     ILIAS_ERROR("mysql-test", ">>> test_join_features FAILED: result size mismatch");
                     co_return {};
                 }
